Edit application structure, replace all to lombok annotations, add logback, create logback config, edit dependencies
