+++ src/main/java/com/bookslib/app/exceptions/BookDaoException.java
@@ -1,0 +1,7 @@
+package com.bookslib.app.exceptions;
+
+public class BookDaoException extends RuntimeException {
+    public BookDaoException(String errorMessage) {
+        super(errorMessage);
+    }
+}
