+++ src/main/java/com/bookslib/app/exceptions/BorrowDaoException.java
@@ -1,0 +1,7 @@
+package com.bookslib.app.exceptions;
+
+public class BorrowDaoException extends ReaderDaoException {
+    public BorrowDaoException(String errorMessage) {
+        super(errorMessage);
+    }
+}
