+++ src/main/java/com/bookslib/app/exceptions/JdbcConnectionException.java
@@ -1,0 +1,7 @@
+package com.bookslib.app.exceptions;
+
+public class JdbcConnectionException extends RuntimeException {
+    public JdbcConnectionException(String errorMessage) {
+        super(errorMessage);
+    }
+}
