+++ src/main/java/com/bookslib/app/exceptions/LibraryServiceException.java
@@ -1,0 +1,7 @@
+package com.bookslib.app.exceptions;
+
+public class LibraryServiceException extends RuntimeException {
+    public LibraryServiceException(String errorMessage) {
+        super(errorMessage);
+    }
+}
