+++ src/main/java/com/bookslib/app/exceptions/ParserServiceException.java
@@ -1,0 +1,7 @@
+package com.bookslib.app.exceptions;
+
+public class ParserServiceException extends RuntimeException {
+    public ParserServiceException(String errorMessage) {
+        super(errorMessage);
+    }
+}
