+++ src/main/java/com/bookslib/app/exceptions/ReaderDaoException.java
@@ -1,0 +1,7 @@
+package com.bookslib.app.exceptions;
+
+public class ReaderDaoException extends RuntimeException {
+    public ReaderDaoException(String errorMessage) {
+        super(errorMessage);
+    }
+}
