--- src/main/java/dao/ReaderDao.java
@@ -1,16 +1,0 @@
-package dao;
-
-import entity.Reader;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface ReaderDao {
-    Reader save(Reader reader);
-
-    List<Reader> findAll();
-
-    Optional<Reader> findById(long readerId);
-
-    List<Reader> findAllByBookId(long bookId);
-}
