--- src/main/java/exceptions/BookDaoException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class BookDaoException extends RuntimeException {
-    public BookDaoException(String errorMessage) {
-        super(errorMessage);
-    }
-}
