--- src/main/java/exceptions/BorrowDaoException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class BorrowDaoException extends ReaderDaoException {
-    public BorrowDaoException(String errorMessage) {
-        super(errorMessage);
-    }
-}
