--- src/main/java/exceptions/JdbcConnectionException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class JdbcConnectionException extends RuntimeException {
-    public JdbcConnectionException(String errorMessage) {
-        super(errorMessage);
-    }
-}
