--- src/main/java/exceptions/LibraryServiceException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class LibraryServiceException extends RuntimeException {
-    public LibraryServiceException(String errorMessage) {
-        super(errorMessage);
-    }
-}
