--- src/main/java/exceptions/ParserServiceException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class ParserServiceException extends RuntimeException {
-    public ParserServiceException(String errorMessage) {
-        super(errorMessage);
-    }
-}
