--- src/main/java/exceptions/ReaderDaoException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class ReaderDaoException extends RuntimeException {
-    public ReaderDaoException(String errorMessage) {
-        super(errorMessage);
-    }
-}
