+++ src/main/resources/application.yml
@@ -5,16 +5,9 @@ spring:
     username: postgres
     password: postgrespw
     driver-class-name: org.postgresql.Driver
-logging:
-  file:
-    name: myapplication.log
-  level:
-    root: INFO
-    sql: INFO
-    web: INFO
-  pattern:
-    console: "%d{dd-MM-yyyy HH:mm:ss.SSS} ([%thread]) %highlight(%-5level)
-      %logger.%M - %msg%n"
-  logback:
-    rollingpolicy:
-      clean-history-on-start: false
+
+server:
+  port: 8081
+
+  servlet:
+    context-path: "/bookslib/api/v1/library"
